Streamlit/requirements.txt
ultralytics
opencv-python-headless
streamlit
Pillow
numpy
torch